Añadir endpoints y operaciones CRUD para las compras

+++ db/definiciones/productos_en_tienda.sql
@@ -1,7 +1,7 @@
 CREATE TABLE productos_en_tienda (
 	id_store BIGINT(20) NOT NULL,
 	id_product BIGINT(20) NOT NULL,
-	price SMALLINT(5) NOT NULL,
+	price DECIMAL(10, 2) NOT NULL,
 	quantity SMALLINT(5) NOT NULL,
 	PRIMARY KEY (id_store, id_product)
 );
